backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
redis
celery
django-celery-beat
gunicorn
